Code cleansed and tested -- Database model fails during estimations

+++ include/ThresholdsBuilderAMS.h
@@ -62,7 +62,7 @@ protected:  // Utils for the class and its kin
 
 	void build_thresholds_vector(const ImportanceFunction& impFun) override;
 
-	void tune(const size_t& numStates,
+	void tune(const uint128_t &numStates,
 			  const size_t& numTrans,
 			  const ImportanceValue& maxImportance,
 			  const unsigned& splitsPerThr) override;
